+++ Extension/src/background/api/filters/common/common-mv2.ts
@@ -165,12 +165,12 @@ export class CommonFilterApi {
      * @param forceRemote Whether to download filter rules from remote resources or
      * from local resources.
      *
-     * **IMPORTANT: `forceRemote` can't be used for MV3** except Quick Fixes
+     * **IMPORTANT: `forceRemote` can't be used for MV3**
      * filter, because we update filters, their metadata, and rulesets with
      * whole extension update, because static ruleset cannot be updated
      * dynamically.
      * To prevent mismatch filters and rulesets version - we do not support load
-     * them from remote, except Quick Fixes Filter, because it will applied
+     * them from remote, because it will applied
      * dynamically.
      *
      * @returns Filter metadata — {@link RegularFilterMetadata}.