+++ tests/helpers/fixtures/getStorageFixtures.ts
@@ -560,21 +560,8 @@ export const getStorageFixturesV7 = (expires: number): StorageData[] => {
             }),
         ).parse(JSON.parse(filtersStateData));
 
-        // Added AdGuard Quick Fixes filter which should be enabled by default.
-        const addedAdGuardQuickFixesFilterId = 24;
-        filtersState[addedAdGuardQuickFixesFilterId] = {
-            // Enabled by default.
-            enabled: true,
-            // Marked as not installed to not remove it as obsoleted
-            // (because after migration it would not have info in metadata).
-            installed: false,
-            // Marked as loaded to update it.
-            loaded: true,
-        };
-
         adgSettings['filters-state'] = JSON.stringify(filtersState);
         settings['adguard-settings'] = adgSettings;
-        settings['raw_filterrules_24.txt'] = '';
         settings['schema-version'] = 7;
 
         return settings;
@@ -590,8 +577,6 @@ export const getStorageFixturesV8 = (expires: number): StorageData[] => {
             settings['schema-version'] = 8;
             return settings;
         }
-
-        settings = removeQuickFixesFilter(settings);
 
         settings['schema-version'] = 8;
 
@@ -625,21 +610,8 @@ export const getStorageFixturesV9 = (expires: number): StorageData[] => {
             }),
         ).parse(JSON.parse(filtersStateData));
 
-        // Added AdGuard Quick Fixes filter which should be enabled by default.
-        const addedAdGuardQuickFixesFilterId = 24;
-        filtersState[addedAdGuardQuickFixesFilterId] = {
-            // Enabled by default.
-            enabled: true,
-            // Marked as not installed to not remove it as obsoleted
-            // (because after migration it would not have info in metadata).
-            installed: false,
-            // Marked as loaded to update it.
-            loaded: true,
-        };
-
         adgSettings['filters-state'] = JSON.stringify(filtersState);
         settings['adguard-settings'] = adgSettings;
-        settings['raw_filterrules_24.txt'] = '';
         settings['schema-version'] = 9;
 
         return settings;
@@ -656,43 +628,10 @@ export const getStorageFixturesV10 = (expires: number): StorageData[] => {
             return settings;
         }
 
-        settings = removeQuickFixesFilter(settings);
-
         settings['schema-version'] = 10;
 
         return settings;
     });
-};
-
-const removeQuickFixesFilter = (settings: StorageData): StorageData => {
-    const adgSettings = settings['adguard-settings'] as any;
-    const filtersStateData = adgSettings['filters-state'];
-
-    if (typeof filtersStateData !== 'string') {
-        throw new Error('Cannot read filters state data');
-    }
-
-    const filtersState = zod.record(
-        zod.string(),
-        zod.object({
-            enabled: zod.boolean(),
-            installed: zod.boolean(),
-            loaded: zod.boolean(),
-        }),
-    ).parse(JSON.parse(filtersStateData));
-
-    // Quick fixes filter was disabled in MV3 to comply with CWR policies.
-    // TODO: remove code totally later.
-
-    // Deprecated AdGuard Quick Fixes filter which should be removed.
-    const deprecatedAdGuardQuickFixesFilterId = 24;
-    delete filtersState[deprecatedAdGuardQuickFixesFilterId];
-
-    adgSettings['filters-state'] = JSON.stringify(filtersState);
-    settings['adguard-settings'] = adgSettings;
-    delete settings['raw_filterrules_24.txt'];
-
-    return settings;
 };
 
 export const getStorageFixturesV11 = (expires: number): StorageData[] => {