+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Network rules with `$important` modifier are applied even if protection is disabled [#3227].
 - Early save reactivation and missing exit prompt during active User Rules/Allowlist save [#3151].
 - Some requests are blocked in 'Inverted allowlist' mode even though there're no websites added to the Allowlist [#3193].
+- Do not enable any filters on settings import if `enabled-filters` is empty [#3136].
 
 ### Changed
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3069]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3069
 [#2934]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2934
 [#3193]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3193
+[#3136]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3136
 
 ## Unreleased v5.2